# docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "EsportHub | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:7271/"
    }
  ],
  "paths": {
...
    "/api/tournaments": {
      "get": {
        "tags": [
          "EsportHub"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultOfIEnumerableOfTournamentResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "EsportHub"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTournamentCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultOfTournamentResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{id}": {
      "get": {
        "tags": [
          "EsportHub"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultOfTournamentResult"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "EsportHub"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTournamentNameCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultOfTournamentResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{id}/group-stage": {
      "post": {
        "tags": [
          "EsportHub"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartGroupStageCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultOfGroupStageResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{id}/group-stage/close": {
      "post": {
        "tags": [
          "EsportHub"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultOfGroupStageResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{id}/group-stage/groups": {
      "get": {
        "tags": [
          "EsportHub"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultOfIEnumerableOfGroupResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{id}/group-stage/groups/{groupId}": {
      "get": {
        "tags": [
          "EsportHub"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultOfGroupResult"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "EsportHub"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupNameCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultOfGroupResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{id}/group-stage/groups/{groupId}/matches": {
      "get": {
        "tags": [
          "EsportHub"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultOfIEnumerableOfGroupMatchResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{id}/group-stage/seed": {
      "post": {
        "tags": [
          "EsportHub"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultOfGroupStageResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{id}/group-stage/groups/{groupId}/matches/{matchId}": {
      "put": {
        "tags": [
          "EsportHub"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "matchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveGroupStageMatchCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultOfGroupMatchResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{id}/knockout-stage": {
      "get": {
        "tags": [
          "EsportHub"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultOfKnockoutStageResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{id}/knockout-stage/matches/{matchId}": {
      "put": {
        "tags": [
          "EsportHub"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "matchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveKnockoutStageMatchCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultOfKnockoutMatchResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{id}/knockout-stage/seed": {
      "post": {
        "tags": [
          "EsportHub"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultOfKnockoutStageResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTournamentCommand": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "GroupResult": {
        "required": [
          "id",
          "name",
          "standings"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "standings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupTeamStandingResult"
            }
          }
        }
      },
      "GroupStageResult": {
        "required": [
          "id",
          "isClosed",
          "groups"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "isClosed": {
            "type": "boolean"
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupSummaryResult"
            }
          }
        }
      },
...
      "KnockoutStageResult": {
        "required": [
          "id",
          "isClosed",
          "matches"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "isClosed": {
            "type": "boolean"
          },
          "matches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KnockoutMatchResult"
            }
          }
        }
      },
...
      "ResolveGroupStageMatchCommand": {
        "required": [
          "tournamentId",
          "groupId",
          "matchId",
          "team1Score",
          "team2Score"
        ],
        "type": "object",
        "properties": {
          "tournamentId": {
            "type": "string",
            "format": "uuid"
          },
          "groupId": {
            "type": "string",
            "format": "uuid"
          },
          "matchId": {
            "type": "string",
            "format": "uuid"
          },
          "team1Score": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "team2Score": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "ResolveKnockoutStageMatchCommand": {
        "required": [
          "tournamentId",
          "matchId",
          "team1Score",
          "team2Score"
        ],
        "type": "object",
        "properties": {
          "tournamentId": {
            "type": "string",
            "format": "uuid"
          },
          "matchId": {
            "type": "string",
            "format": "uuid"
          },
          "team1Score": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "team2Score": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "ResultOfGroupMatchResult": {
        "required": [
          "value"
        ],
        "type": "object",
        "properties": {
          "value": {
            "$ref": "#/components/schemas/GroupMatchResult"
          },
          "status": {
            "$ref": "#/components/schemas/ResultStatus"
          },
          "isSuccess": {
            "type": "boolean"
          },
          "successMessage": {
            "type": [
              "null",
              "string"
            ]
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ]
          },
          "location": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "validationErrors": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "ResultOfGroupResult": {
        "required": [
          "value"
        ],
        "type": "object",
        "properties": {
          "value": {
            "$ref": "#/components/schemas/GroupResult"
          },
          "status": {
            "$ref": "#/components/schemas/ResultStatus"
          },
          "isSuccess": {
            "type": "boolean"
          },
          "successMessage": {
            "type": [
              "null",
              "string"
            ]
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ]
          },
          "location": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "validationErrors": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "ResultOfGroupStageResult": {
        "required": [
          "value"
        ],
        "type": "object",
        "properties": {
          "value": {
            "$ref": "#/components/schemas/GroupStageResult"
          },
          "status": {
            "$ref": "#/components/schemas/ResultStatus"
          },
          "isSuccess": {
            "type": "boolean"
          },
          "successMessage": {
            "type": [
              "null",
              "string"
            ]
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ]
          },
          "location": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "validationErrors": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "ResultOfIEnumerableOfGroupMatchResult": {
        "required": [
          "value"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/GroupMatchResult"
            }
          },
          "status": {
            "$ref": "#/components/schemas/ResultStatus"
          },
          "isSuccess": {
            "type": "boolean"
          },
          "successMessage": {
            "type": [
              "null",
              "string"
            ]
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ]
          },
          "location": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "validationErrors": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "ResultOfIEnumerableOfGroupResult": {
        "required": [
          "value"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/GroupResult"
            }
          },
          "status": {
            "$ref": "#/components/schemas/ResultStatus"
          },
          "isSuccess": {
            "type": "boolean"
          },
          "successMessage": {
            "type": [
              "null",
              "string"
            ]
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ]
          },
          "location": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "validationErrors": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "ResultOfIEnumerableOfTournamentResult": {
        "required": [
          "value"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/TournamentResult"
            }
          },
          "status": {
            "$ref": "#/components/schemas/ResultStatus"
          },
          "isSuccess": {
            "type": "boolean"
          },
          "successMessage": {
            "type": [
              "null",
              "string"
            ]
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ]
          },
          "location": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "validationErrors": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "ResultOfKnockoutMatchResult": {
        "required": [
          "value"
        ],
        "type": "object",
        "properties": {
          "value": {
            "$ref": "#/components/schemas/KnockoutMatchResult"
          },
          "status": {
            "$ref": "#/components/schemas/ResultStatus"
          },
          "isSuccess": {
            "type": "boolean"
          },
          "successMessage": {
            "type": [
              "null",
              "string"
            ]
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ]
          },
          "location": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "validationErrors": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "ResultOfKnockoutStageResult": {
        "required": [
          "value"
        ],
        "type": "object",
        "properties": {
          "value": {
            "$ref": "#/components/schemas/KnockoutStageResult"
          },
          "status": {
            "$ref": "#/components/schemas/ResultStatus"
          },
          "isSuccess": {
            "type": "boolean"
          },
          "successMessage": {
            "type": [
              "null",
              "string"
            ]
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ]
          },
          "location": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "validationErrors": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "ResultOfTournamentResult": {
        "required": [
          "value"
        ],
        "type": "object",
        "properties": {
          "value": {
            "$ref": "#/components/schemas/TournamentResult"
          },
          "status": {
            "$ref": "#/components/schemas/ResultStatus"
          },
          "isSuccess": {
            "type": "boolean"
          },
          "successMessage": {
            "type": [
              "null",
              "string"
            ]
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ]
          },
          "location": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "validationErrors": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "ResultStatus": {
        "type": "integer"
      },
...
      "StartGroupStageCommand": {
        "required": [
          "tournamentId",
          "groupNames"
        ],
        "type": "object",
        "properties": {
          "tournamentId": {
            "type": "string",
            "format": "uuid"
          },
          "groupNames": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "TournamentResult": {
        "required": [
          "id",
          "name",
          "startDate",
          "status",
          "winner"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "startDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/TournamentStatus"
          },
          "winner": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "TournamentStatus": {
        "type": "integer"
      },
...
      "UpdateGroupNameCommand": {
        "required": [
          "tournamentId",
          "groupId",
          "name"
        ],
        "type": "object",
        "properties": {
          "tournamentId": {
            "type": "string",
            "format": "uuid"
          },
          "groupId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "UpdateTournamentNameCommand": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": [
              "null",
              "string"
            ]
          },
          "errorMessage": {
            "type": [
              "null",
              "string"
            ]
          },
          "errorCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "severity": {
            "$ref": "#/components/schemas/ValidationSeverity"
          }
        }
      },
      "ValidationSeverity": {
        "type": "integer"
      }
    }
  },
...
}